MAJOR UPDATE
-new bluenoise
-ACTUAL FIX for FRACT OFFSET
-deprecated textureshaders in favor of fractional offset working properly.
-distance to lighting is now voxel accurate, removed alot of 0.5f's
-improved singletons with constinit usage - no more "hidden" guards around major class singletons.
-volumetric changes to be comptabile and perfectly aligned with the fractional offset.
-the fractional offset is actually working properly, outstanding issue for 1.5yrs - nice smooth camera scrolling w/o jittered world objects is nice.
-`other adds and minor changes

+++ MinCity/cPoliceCarGameObject.h
@@ -31,8 +31,8 @@ namespace world
 
 		static constexpr fp_seconds const
 			LIGHT_SWITCH_INTERVAL = duration_cast<fp_seconds>(milliseconds(75));
-		static constexpr float const
-			INV_LIGHT_SWITCH_INTERVAL = 1.0f / LIGHT_SWITCH_INTERVAL.count();
+		static constexpr double const
+			INV_LIGHT_SWITCH_INTERVAL = 1.0 / LIGHT_SWITCH_INTERVAL.count();
 	public:
 		constexpr virtual types::game_object_t const to_type() const override final {
 			return(types::game_object_t::PoliceCarGameObject);
